requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
amqp==5.2.0
appdirs==1.4.3
apturl==0.5.2
asgiref==3.7.2
astroid==2.3.3
async-timeout==4.0.3
attrs==19.3.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.12.2
billiard==4.2.0
blinker==1.7.0
Brlapi==0.7.0
Brotli==1.1.0
cached-property==1.5.1
cachetools==5.3.2
celery==5.3.6
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.3
command-not-found==0.3
cryptography==41.0.1
cssselect2==0.7.0
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
Deprecated==1.2.14
distlib==0.3.0
distro==1.4.0
distro-info==0.23+ubuntu1.1
Django==4.2.5
django-admin-list-filter-dropdown==1.0.3
django-admin-rangefilter==0.10.0
django-appconf==1.0.5
django-ckeditor==6.6.1
django-cors-headers==4.1.0
django-jazzmin==2.6.0
django-js-asset==2.1.0
django-login-required-middleware==0.9.0
django-select2==8.1.2
django-storages==1.13.2
django-tinymce==3.6.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
drf-user-activity-tracker==1.0.1
duplicity==0.8.12.0
entrypoints==0.3
et-xmlfile==1.1.0
fasteners==0.14.1
feedgenerator==2.1.0
filelock==3.0.12
flask==3.0.0
fonttools==4.42.1
frozenlist==1.4.0
future==0.18.2
google-api-core==2.14.0
google-api-python-client==2.107.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
googleapis-common-protos==1.61.0
gunicorn==21.2.0
html5lib==1.1
httplib2==0.22.0
idna==2.8
importlib-metadata==6.8.0
isort==4.3.21
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==3.2.0
keyring==18.0.1
kombu==5.3.4
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
load-dotenv==0.1.0
lockfile==0.12.2
louis==3.12.0
Mako==1.1.0
MarkupSafe==2.1.3
mccabe==0.6.1
monotonic==1.5
more-itertools==4.2.0
multidict==6.0.4
mutagen==1.47.0
mysqlclient==2.1.1
netifaces==0.10.4
numpy==1.24.4
oauthlib==3.1.0
olefile==0.46
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
paramiko==2.6.0
pdfkit==1.0.0
pexpect==4.6.0
Pillow==10.0.0
prompt-toolkit==3.0.41
protobuf==4.25.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.16.2
pycparser==2.21
pycryptodomex==3.19.0
pycups==1.9.73
pydyf==0.7.0
PyGObject==3.36.0
PyJWT==2.4.0
PyNaCl==1.3.0
pyparsing==3.1.1
pyphen==0.14.0
pypng==0.20220715.0
pyrsistent==0.15.5
pySmartDL==1.3.4
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.2
python-debian==0.1.36+ubuntu1.1
python-dotenv==0.20.0
pytube==15.0.0
pytz==2022.7.1
pyxdg==0.26
PyYAML==5.3.1
qrcode==7.4.2
redis==4.1.4
requests==2.31.0
requests-unixsocket==0.2.0
rest-condition==1.0.3
rest-framework-simplejwt==0.0.2
rsa==4.9
schedule==1.2.1
SecretStorage==2.3.1
simplejson==3.16.0
six==1.13.0
soupsieve==2.5
sqlparse==0.4.4
sslcommerz-python==0.0.7
systemd-python==234
texttable==1.6.2
tinycss2==1.2.1
typed-ast==1.4.0
typing-extensions==4.8.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==2.1.0
usb-creator==0.3.7
vine==5.1.0
virtualenv==20.0.17
wadllib==1.3.3
wcwidth==0.2.12
weasyprint==59.0
webencodings==0.5.1
websocket-client==0.53.0
websockets==12.0
werkzeug==3.0.1
wrapt==1.11.2
xkit==0.0.0
yarl==1.9.2
youtube-dl==2021.12.17
yt-dlp==2023.11.16
zipp==1.0.0
zopfli==0.2.3
